Frontend/requirements.txt
streamlit
requests
google-cloud-monitoring
google-cloud-compute
psutil
